Python/Practice Projects/expensetrack/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Security-Too
Flask-JWT-Extended
psycopg2-binary
python-dotenv
